fix(audit): corrige 7 bugs criticos e medios detectados na auditoria profunda
BUG-01: routers.py importava guardian_service de health (watchdog MT5) em vez de application.services (risk guardian) -> /api/state crashava em runtime
BUG-02: arbiter_service.py chamava event_bus.publish sem await em 3 pontos e passava str em vez de BaseEvent -> todos os eventos de rejeicao eram perdidos silenciosamente
BUG-04: engine.py tinha importacao duplicada de arbiter_service
BUG-05: guardian_service.get_state() sempre retornava daily_drawdown=0.0. Adicionada task periodica que le todays_realized_pnl do event_store
BUG-07: get_kelly_volume() dividia risk_pct por 100 duas vezes -> volume Kelly era 100x menor que esperado
BUG-09: Substituido .dict() deprecated (Pydantic v2) por .model_dump()
BUG-11: ScanConfig.interval_sec mudado de int para float para aceitar valores decimais da UI

+++ src/application/services/engine_config_service.py
@@ -35,7 +35,7 @@ class TieBreakConfig(BaseModel):
 
 class ScanConfig(BaseModel):
     """Strategy Engine scan settings."""
-    interval_sec: int = 60           # Intervalo entre ciclos de análise (segundos)
+    interval_sec: float = 60           # Intervalo entre ciclos de análise (segundos)
     symbol_delay_sec: float = 0.5    # Delay entre análise de cada símbolo (throttle MT5)
